Disable multi floating ranges when using Neutron
Currently fuel-library doesn't seem to support multiple
floating ranges when using Neutron. Neutron it self does
support multiple floating ranges when using the Python api
although the CLI has no support for this.

This patch will first disable it in the UI as a first step
and the actual support for multiple floating ip ranges with Neutron
in fuel will be added later. We should re-enable the functionality
in UI as soon as we have support in Fuel.

Hacky way - needs change after advanced networking implementation

Related-bug: #1371363

Change-Id: I7917f5b8419b4f36e80a9d1a59291d67c832a4bd