Remove unneeded cluster networking functional tests
This change removes test case when erroneous response is expected from
Nailgun network configuration handler in case range with only one
element has been submitted for public network.

The reason it is done because in neareast future change for Nailgun
which disables allocation of VIPs on PUT operation for the handler (
and thus corresponding error is not raised) will be delivered and the
case is no more relevant.

Change-Id: If21b789dbb798dd1597dc03903d333d7da6648d7
Blueprint: allow-any-vip