Enable all ES6 features for the whole project
Some of ES6 features (like template strings, let, const, etc)
were disabled as we couldn't enable them in our functional
tests, so we would end up with 2 code styles. Since the tests
are also transpiled, we can enable all the features.

no-use-before-define was disabled as it started to cause errors
on the networks tab on the code which looked totally fine, and
fixing them would cause the code look worse.

Change-Id: I5e20b1f7ebb6c18e31faa19271b73f53b5aea878